solo puede eliminar adjuntos si el report no está validado

+++ protected/controllers/RCamionPropioController.php
@@ -265,17 +265,18 @@ class RCamionPropioController extends Controller
 							}  
 						}
 					}
-					if(isset($_POST['eliminar'])){
-						$eliminables = $_POST['eliminar'];
-						if(isset($eliminables)){
-							foreach($eliminables as $archivo => $status){
-								if($status == 'on'){
-									unlink($path . DIRECTORY_SEPARATOR . $archivo);
+					if($model->validado == 0){
+						if(isset($_POST['eliminar'])){
+							$eliminables = $_POST['eliminar'];
+							if(isset($eliminables)){
+								foreach($eliminables as $archivo => $status){
+									if($status == 'on'){
+										unlink($path . DIRECTORY_SEPARATOR . $archivo);
+									}
 								}
 							}
 						}
 					}
-
 					//end archivos del report
 
 					$valid = true;
