+++ protected/controllers/REquipoArrendadoController.php
@@ -261,12 +261,14 @@ class REquipoArrendadoController extends Controller
 							}  
 						}
 					}
-					if(isset($_POST['eliminar'])){
-						$eliminables = $_POST['eliminar'];
-						if(isset($eliminables)){
-							foreach($eliminables as $archivo => $status){
-								if($status == 'on'){
-									unlink($path . DIRECTORY_SEPARATOR . $archivo);
+					if($model->validado == 0){
+						if(isset($_POST['eliminar'])){
+							$eliminables = $_POST['eliminar'];
+							if(isset($eliminables)){
+								foreach($eliminables as $archivo => $status){
+									if($status == 'on'){
+										unlink($path . DIRECTORY_SEPARATOR . $archivo);
+									}
 								}
 							}
 						}
